# core-banking-psdx-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-psdx
  description: PSD2/PSD3 and FIDA Regulatory Compliance Service
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/psdx
    description: Development Environment
  - url: /
    description: Local Development Environment
security:
  - apiKey: []
  - bearer: []
...
paths:
...
  /api/v1/payments/{paymentId}/authorize:
    post:
      tags:
        - Payment Initiation Services
      summary: Authorize payment
      description: |
        ## Authorize a Payment

        This endpoint authorizes a payment using Strong Customer Authentication (SCA).

        ### Description
        This operation is part of the Payment Initiation Service (PIS) under PSD2/PSD3 regulations.
        It completes the authorization process for a payment by applying the authentication code
        obtained through the Strong Customer Authentication (SCA) process.

        ### Request Headers
        * `X-Consent-ID` - The unique identifier of the consent that authorizes access to this payment

        ### Path Parameters
        * `paymentId` - The unique identifier of the payment to authorize

        ### Request Body
        * `authorizationCode` - The authorization code obtained from the SCA process

        ### Processing
        The system will:
        1. Validate the consent ID provided in the header
        2. Check that the consent is valid, active, and covers the specified payment
        3. Verify that the payment exists and is associated with the consent
        4. Check if the payment is in a state that allows authorization
        5. Validate the authorization code against the SCA challenge
        6. Update the payment status to "AUTHORIZED"
        7. Initiate the payment execution process

        ### Response
        Returns the updated payment object with status changed to "AUTHORIZED".

        ### Authorization Flow
        1. Payment is initiated (status: RECEIVED)
        2. SCA is triggered (status: PENDING)
        3. Customer completes authentication and receives an authorization code
        4. TPP submits the authorization code to this endpoint
        5. Payment is authorized (status: AUTHORIZED)
        6. Payment is executed (status will change to EXECUTED asynchronously)

        ### SCA Requirements
        * Strong Customer Authentication is mandatory for payment authorization under PSD2/PSD3
        * SCA typically involves two-factor authentication (2FA)
        * The authorization code has a limited validity period
        * Multiple failed authorization attempts may lock the payment

        ### Error Scenarios
        * If the payment doesn't exist, returns a 404 Not Found response
        * If the payment exists but is not covered by the consent, returns a 403 Forbidden response
        * If the consent is invalid or expired, returns a 401 Unauthorized response
        * If the authorization code is invalid or expired, returns a 401 Unauthorized response
        * If the payment cannot be authorized (e.g., already executed), returns a 400 Bad Request response

        ### Regulatory Compliance
        * This endpoint implements the Payment Authorization Service as defined in PSD2/PSD3
        * Strong Customer Authentication is a regulatory requirement
        * Access is logged for regulatory reporting and audit purposes

        ### Security Considerations
        * All requests must include a valid consent ID
        * Authorization codes are single-use and time-limited
        * Failed authorization attempts are monitored for fraud prevention
        * All authorization attempts are logged for audit purposes
      operationId: authorizePayment
      parameters:
        - name: X-Consent-ID
          in: header
          description: ID of the consent
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentId
          in: path
          description: ID of the payment
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: PSU-ID
          in: header
          description: ID of the customer
          schema:
            type: string
          example: user@example.com
        - name: PSU-IP-Address
          in: header
          description: IP address of the customer
          schema:
            type: string
          example: 192.168.1.1
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: Authorization code from SCA
        required: true
      responses:
        '200':
          description: Payment authorized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentDTO'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentDTO'
...
  /api/v1/payments/{paymentId}/status:
    get:
      tags:
        - Payment Initiation Services
      summary: Get payment status
      description: |
        ## Retrieve Payment Status

        This endpoint retrieves the current status of a specific payment.

        ### Description
        This operation is part of the Payment Initiation Service (PIS) under PSD2/PSD3 regulations.
        It allows Third Party Providers (TPPs) to check the status of a payment that was previously
        initiated through the API.

        ### Request Headers
        * `X-Consent-ID` - The unique identifier of the consent that authorizes access to this payment

        ### Path Parameters
        * `paymentId` - The unique identifier of the payment whose status is being retrieved

        ### Processing
        The system will:
        1. Validate the consent ID provided in the header
        2. Check that the consent is valid, active, and covers the specified payment
        3. Verify that the payment exists and is associated with the consent
        4. Retrieve the current status of the payment

        ### Response
        Returns a payment status object containing:
        * Payment identifier
        * Current status (e.g., "RECEIVED", "PENDING", "AUTHORIZED", "EXECUTED", "REJECTED", "CANCELLED")
        * Status timestamp (when the status was last updated)
        * Status reason (if applicable, especially for rejected payments)
        * Additional status details (if available)

        ### Status Definitions
        * **RECEIVED** - Payment has been received but not yet processed
        * **PENDING** - Payment is awaiting authorization or processing
        * **AUTHORIZED** - Payment has been authorized but not yet executed
        * **EXECUTED** - Payment has been successfully executed
        * **REJECTED** - Payment has been rejected (with reason)
        * **CANCELLED** - Payment has been cancelled by the user or system

        ### Error Scenarios
        * If the payment doesn't exist, returns a 404 Not Found response
        * If the payment exists but is not covered by the consent, returns a 403 Forbidden response
        * If the consent is invalid or expired, returns a 401 Unauthorized response

        ### Regulatory Compliance
        * This endpoint implements the Payment Status Check as defined in PSD2/PSD3
        * Access is logged for regulatory reporting and audit purposes

        ### Security Considerations
        * All requests must include a valid consent ID
        * Consent validation includes checking expiration and revocation status
        * Status checks may be rate-limited to prevent excessive polling
      operationId: getPaymentStatus
      parameters:
        - name: X-Consent-ID
          in: header
          description: ID of the consent
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentId
          in: path
          description: ID of the payment
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Request-ID
          in: header
          description: Unique ID for the request
          schema:
            type: string
          example: 99391c7e-ad88-49ec-a2ad-99ddcb1f7721
        - name: PSU-ID
          in: header
          description: ID of the customer
          schema:
            type: string
          example: user@example.com
        - name: PSU-IP-Address
          in: header
          description: IP address of the customer
          schema:
            type: string
          example: 192.168.1.1
      responses:
        '200':
          description: Payment status found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentStatusDTO'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentStatusDTO'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentStatusDTO'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentStatusDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PSDPaymentStatusDTO'
...
components:
  schemas:
...
    PSDPaymentDTO:
      required:
        - consentId
        - instructedAmount
        - paymentId
      type: object
      properties:
        paymentId:
          type: string
          description: Unique identifier of the payment
          format: uuid
        endToEndIdentification:
          maxLength: 35
          minLength: 0
          pattern: ^[A-Za-z0-9/\-?:().,'+\s]*$
          type: string
          description: End-to-end identifier
          example: E2E-ID-123
        consentId:
          type: string
          description: ID of the consent used for the payment
          format: uuid
        paymentType:
          pattern: ^(sepa-credit-transfers|instant-sepa-credit-transfers|target-2-payments|cross-border-credit-transfers)$
          type: string
          description: Type of the payment
          example: sepa-credit-transfers
        transactionStatus:
          pattern: ^(ACCP|ACSC|ACSP|ACTC|ACWC|ACWP|PDNG|RJCT|CANC)$
          type: string
          description: Status of the payment
          example: ACCP
        debtorAccount:
          $ref: '#/components/schemas/PSDAccountReferenceDTO'
        creditorName:
          maxLength: 70
          minLength: 0
          type: string
          description: Creditor name
          example: John Doe
        creditorAccount:
          $ref: '#/components/schemas/PSDAccountReferenceDTO'
        creditorAddress:
          $ref: '#/components/schemas/PSDAddressDTO'
        instructedAmount:
          $ref: '#/components/schemas/PSDAmountDTO'
        remittanceInformationUnstructured:
          type: string
          description: Remittance information unstructured
          example: Invoice 123
        remittanceInformationStructured:
          $ref: '#/components/schemas/PSDRemittanceDTO'
        requestedExecutionDate:
          type: string
          description: Requested execution date
          format: date
        get_links:
          $ref: '#/components/schemas/PSDLinksDTO'
        createdAt:
          type: string
          description: Date and time when the payment was created
          format: date-time
        updatedAt:
          type: string
          description: Date and time when the payment was last updated
          format: date-time
      description: PSD Payment information
...
  securitySchemes:
    apiKey:
      type: apiKey
      description: API key for TPP authentication
      name: X-API-KEY
      in: header
    bearer:
      type: http
      description: JWT token for user authentication
      scheme: bearer
      bearerFormat: JWT